[moe] add mixtral moe layer
